main screen refactoring
